conversions

+++ Opcodes/serial.c
@@ -395,7 +395,7 @@ int32_t serialWrite_S(CSOUND *csound, SERIALWRITE *p)
     if (UNLIKELY(write((int32_t)*p->port,
                        ((STRINGDAT*)p->toWrite)->data,
                        ((STRINGDAT*)p->toWrite)->size))!=
-        ((STRINGDAT*)p->toWrite)->size) /* Does Windows write behave correctly? */
+        (long)((STRINGDAT*)p->toWrite)->size) /* Does Windows write behave correctly? */
         return NOTOK;
 #else
       int32_t nbytes;
